v1.2.2
Added an introductory message when a player logs in and hasn't used the
plugin before, if they have the permission to use the plugin. Allowed
use of the /AdminSaveFile command from sources other than players. Made
the /AdminSaveRam command more convenient by not requiring "overwrite"
at the end of the command unless the snapshot does exist.