v1.2.3
Remastered the commands so they make more sense and are easier to
remember. Added the ability to detect failures in loading and to not
save on disable as a result of said failures. More small performance
tweaks as well. And probably some other stuff I forgot since I've worked
on this one so long.